Refactor upload collection/default row handling to share status handling.
Removes some divergence.
